added interfaces and controllers in external dir
