+++ src/raccoon/external/controllers/__init__.py
@@ -1,0 +1,4 @@
+from .github import GitHubController
+from .bitbucket import BitbucketController
+from .jenkins import JenkinsController
+from .salt import SaltController
