+++ src/raccoon/external/controllers/jenkins.py
@@ -1,13 +1,13 @@
 import logging
 
-from tornado import gen
 from mongoengine.errors import DoesNotExist
+from tornado import gen
 
-from . import BaseController
 from ..interfaces.jenkins import JenkinsInterface
-from ..models import Flow, Connector, Job
-from ..utils.decorators import authenticated
-from ..utils.exceptions import ReplyError
+from ...controllers import BaseController
+from ...models import Flow, Connector, Job
+from ...utils.decorators import authenticated
+from ...utils.exceptions import ReplyError
 
 log = logging.getLogger(__name__)
