+++ src/raccoon/external/controllers/salt.py
@@ -1,13 +1,13 @@
 import logging
 
-from tornado import gen
 from mongoengine.errors import DoesNotExist
+from tornado import gen
 
-from . import BaseController
 from ..interfaces.salt import SaltStackInterface
-from ..models import Connector, AuditLog
-from ..utils.decorators import authenticated
-from ..utils.exceptions import ReplyError
+from ...controllers import BaseController
+from ...models import Connector, AuditLog
+from ...utils.decorators import authenticated
+from ...utils.exceptions import ReplyError
 
 
 log = logging.getLogger(__name__)
