+++ src/raccoon/external/interfaces/base.py
@@ -4,7 +4,7 @@ import logging
 from tornado import gen
 from tornado.httpclient import AsyncHTTPClient, HTTPRequest, HTTPError
 
-from ..utils.exceptions import ReplyError
+from ...utils.exceptions import ReplyError
 
 log = logging.getLogger(__name__)
