+++ src/raccoon/external/interfaces/jenkins.py
@@ -6,11 +6,11 @@ from tornado import gen
 from mongoengine.errors import DoesNotExist
 
 from . import BaseInterface, REGISTERED
-from ..models import Task, Build, Project, Environment, Install, AuditLog
-from ..utils.exceptions import ReplyError
-from ..utils.request import broadcast
-from ..utils.utils import translate_job_arguments
-from ..tasks.jenkins import JenkinsQueueWatcherTask, PENDING
+from ...models import Task, Build, Project, Environment, Install, AuditLog
+from ...utils.exceptions import ReplyError
+from ...utils.request import broadcast
+from ...utils.utils import translate_job_arguments
+from ...tasks.jenkins import JenkinsQueueWatcherTask, PENDING
 
 
 log = logging.getLogger(__name__)
