+++ src/raccoon/urls.py
@@ -2,6 +2,7 @@ import re
 import logging
 
 from .controllers import *
+from .external.controllers import *
 from .utils.exceptions import ReplyError
 
 log = logging.getLogger(__name__)
